test/bpf: Move ipv6/lpm tests into bpf/
Follow the golang approach of placing test code at the location
following the pattern of substitution %s.c -> %s_test.c.
